magic_commit/tests/test_config.py
# TODO: Fix this test file
# https://github.com/heyodai/magic-commit/issues/17

# from ..magic_commit import get_api_key, set_api_key, get_model, set_model
# import unittest
# from unittest.mock import patch, mock_open, MagicMock
# class TestMagicCommitConfig(unittest.TestCase):

#     @patch('.configparser.ConfigParser')
#     def test_get_api_key(self, mock_config_parser):
#         # Setup the mock to return a predefined API key
#         mock_config_parser.return_value.get.return_value = "test_api_key"

#         # Call the function to test
#         api_key = get_api_key()

#         # Check if the correct API key is returned
#         self.assertEqual(api_key, "test_api_key")

#     @patch('.configparser.ConfigParser')
#     @patch('.open', new_callable=mock_open)
#     def test_set_api_key(self, mock_file, mock_config_parser):
#         # Attempt to set an API key and ensure the file write is called
#         set_api_key("new_test_api_key")

#         # Ensure that the config file write method was called
#         mock_file.assert_called_with(CONFIG_FILE, "w")
#         mock_config_parser.return_value.write.assert_called()

#     @patch('.os.path.exists')
#     @patch('.configparser.ConfigParser')
#     @patch('.open', new_callable=mock_open, read_data="[DEFAULT]\napi_key=test")
#     def test_get_api_key_with_no_key_set(self, mock_file, mock_config_parser, mock_path_exists):
#         # Setup to simulate no API key set
#         mock_path_exists.return_value = False
#         mock_config_parser.return_value.get.side_effect = KeyError("no key")

#         # Call the function
#         api_key = get_api_key()

#         # Check if the function handles the absence of an API key correctly
#         self.assertIsNone(api_key)

#     @patch('.Logger')
#     def test_set_model(self, mock_logger):
#         # This test is a placeholder, the actual set_model function is not implemented
#         set_model("gpt-3.5-turbo", mock_logger)
#         # You would add assertions here based on the actual implementation

#     def test_get_model(self):
#         # Since get_model is not implemented, this test simply checks the placeholder return value
#         self.assertEqual(get_model(), "gpt-3.5-turbo")

# # This allows the tests to be run from the command line
# if __name__ == '__main__':
#     unittest.main()
